+++ projects/citizen/src/app/core/components/approval-page/approval-page.component.ts
@@ -34,5 +34,8 @@ export class ApprovalPageComponent {
   public notFoundDetail = '';
 
   @Input()
+  public notFoundReasons: string[] = [];
+
+  @Input()
   public loading = true;
 }