+++ projects/citizen/src/app/core/exceptions.ts
@@ -6,8 +6,3 @@
 
 export const insufficientAcrException = 'InsufficientAcrException';
 export const emailDoesNotMatchException = 'EmailDoesNotMatchException';
-export const collectionAlreadySignedException = 'CollectionAlreadySignedException';
-export const decreeAlreadySignedException = 'DecreeAlreadySignedException';
-export const collectionMaxElectronicSignatureCountReachedException = 'CollectionMaxElectronicSignatureCountReachedException';
-export const decreeMaxElectronicSignatureCountReachedException = 'DecreeMaxElectronicSignatureCountReachedException';
-export const personOrVotingRightNotFoundException = 'PersonOrVotingRightNotFoundException';