+++ projects/citizen/src/app/core/models/referendum.model.ts
@@ -7,8 +7,9 @@
 import {
   ListMyReferendumsResponse as ListMyReferendumsResponseProto,
   Referendum as ReferendumProto,
+  SimpleDecree as SimpleDecreeProto,
 } from '@abraxas/voting-ecollecting-proto/citizen';
-import { Referendum as ReferendumShared, SimpleDecree as SimpleDecreeShared } from 'ecollecting-lib';
+import { fromProtoDate, Referendum as ReferendumShared, SimpleDecree as SimpleDecreeShared } from 'ecollecting-lib';
 import { Collection, mapCollectionToModel } from './collection.model';
 import { Decree, mapDecreeToModel } from './decree.model';
 import { CollectionCameNotAboutReason, CollectionPeriodState, CollectionType } from '@abraxas/voting-ecollecting-proto';
@@ -30,6 +31,9 @@ export interface ListMyReferendumsResponse extends Omit<ListMyReferendumsRespons
 export interface SimpleDecree extends SimpleDecreeShared {
   cameNotAboutReason?: CollectionCameNotAboutReason;
   periodState?: CollectionPeriodState;
+  collectionStartDate?: Date;
+  collectionEndDate?: Date;
+  signatureListSubmissionEndDate?: Date;
 }
 
 export function mapToListMyReferendumsResponse(response: ListMyReferendumsResponseProto): ListMyReferendumsResponse {
@@ -41,10 +45,21 @@ export function mapToListMyReferendumsResponse(response: ListMyReferendumsRespon
 
 export function mapReferendumToModel(referendumProto: ReferendumProto): Referendum {
   const collection = mapCollectionToModel(referendumProto.collection!);
+  const decree = referendumProto.decree ? mapSimpleDecreeToModel(referendumProto.decree) : undefined;
   return {
     ...referendumProto.toObject(),
     collection: collection,
+    decree: decree,
   } as Referendum;
+}
+
+export function mapSimpleDecreeToModel(simpleDecreeProto: SimpleDecreeProto): SimpleDecree {
+  return {
+    ...simpleDecreeProto.toObject(),
+    collectionStartDate: fromProtoDate(simpleDecreeProto.collectionStartDate),
+    collectionEndDate: fromProtoDate(simpleDecreeProto.collectionEndDate),
+    signatureListSubmissionEndDate: fromProtoDate(simpleDecreeProto.signatureListSubmissionEndDate),
+  } as SimpleDecree;
 }
 
 export function isReferendum(obj?: Initiative | Referendum): obj is Referendum {