+++ projects/citizen/src/app/user/user-seek-referendum/seek-referendum-detail/seek-referendum-detail-overview/seek-referendum-detail-overview.component.ts
@@ -4,14 +4,36 @@
  * For license information see LICENSE file.
  */
 
-import { Component } from '@angular/core';
-import { TranslateDirective } from '@ngx-translate/core';
-import { CardModule } from '@abraxas/base-components';
+import { Component, inject, OnDestroy } from '@angular/core';
+import { TranslateDirective, TranslatePipe } from '@ngx-translate/core';
+import { AlertBarModule, CardModule } from '@abraxas/base-components';
+import { ActivatedRoute } from '@angular/router';
+import { Subscription } from 'rxjs';
+import { Referendum } from '../../../../core/models/referendum.model';
+import { CollectionPeriodState, CollectionState } from '@abraxas/voting-ecollecting-proto';
 
 @Component({
   selector: 'app-seek-referendum-detail-overview',
   templateUrl: './seek-referendum-detail-overview.component.html',
   styleUrl: 'seek-referendum-detail-overview.component.scss',
-  imports: [TranslateDirective, CardModule],
+  imports: [TranslateDirective, CardModule, AlertBarModule, TranslatePipe],
 })
-export class SeekReferendumDetailOverviewComponent {}
+export class SeekReferendumDetailOverviewComponent implements OnDestroy {
+  protected collectionStates = CollectionState;
+  protected collectionPeriodStates = CollectionPeriodState;
+  protected referendum?: Referendum;
+
+  private routeSubscription: Subscription;
+
+  constructor() {
+    const route = inject(ActivatedRoute);
+
+    this.routeSubscription = route.parent!.data.subscribe(async ({ referendum }) => {
+      this.referendum = referendum;
+    });
+  }
+
+  public ngOnDestroy(): void {
+    this.routeSubscription.unsubscribe();
+  }
+}