+++ projects/ecollecting-lib/src/lib/core/utils/error.utils.ts
@@ -5,12 +5,9 @@
  */
 
 import { GrpcStatusEvent } from '@ngx-grpc/common';
+import { isErrorType } from '@abraxas/voting-lib';
 
 export function getGrpcErrorOrThrow(e: any, errTypes: string[]): string {
-  if (!(e instanceof GrpcStatusEvent)) {
-    throw e;
-  }
-
   for (const errType of errTypes) {
     if (isGrpcError(e, errType)) {
       return errType;
@@ -21,10 +18,8 @@ export function getGrpcErrorOrThrow(e: any, errTypes: string[]): string {
 }
 
 export function isGrpcError(e: any, errType: string): boolean {
-  return (
-    e instanceof GrpcStatusEvent &&
-    (e.statusMessage === errType || e.statusMessage.startsWith(errType + ':') || e.statusMessage.startsWith(errType + ' '))
-  );
+  // the grpc errors get mapped to the lib error in the interceptor
+  return isErrorType(e, errType);
 }
 
 export function isGrpcNotFoundError(e: any): boolean {